PROD-230: Create a new ReportComponent to hold the components related to reports.
* Registered this component with 'report' selector.
* The HTML template of this component is empty at this point.
* Bootstrapped this component in the main.ts file so that it is accessible.

PROD-230: Add bootstrap cdn in the styleUrls to use it in our HTML template.

PROD-230: Write a empty ReportHandler for handling the GET requests made by the frontend
to fetch the reports.

* Also, indented the code properly to confirm the python standards.

PROD-230: Add basic skeleton file(more room for improvement in future) report.service which
focuses on getting the data required for the report.

Fix-me! Inject firebase in ReportComponent so that we can use it to get the UID.